modify: change delegateforwarder to specific implementation.
